refactor: [#200] reorganize E2E testing modules
- Move process_runner.rs to src/testing/e2e/ (remove black_box wrapper)
- Move tasks/verify_dependencies.rs to src/testing/e2e/tasks/black_box/
- Remove empty black_box directory
- Update all imports and re-exports
