+++ src/testing/e2e/tasks/black_box/mod.rs
